More than a calculator
This scientific calculator has many capabilities to handle scientific operations other than simple calculations with elegant design. It comes with light and dark theme which is decided based on the local time. Currently this is at its earlier stage so some of the functions may or may not work properly. You are welcome to give it a try.